+++ admin/src/domains.js
@@ -1,40 +1,19 @@
 import React from 'react';
-import { CardActions } from 'material-ui/Card';
 import {
-    Create, CreateButton, Datagrid, DateField, DeleteButton, DisabledInput, Edit, EditButton, Filter, List, ListButton, RefreshButton, required, Show, ShowButton, SimpleForm,
-    SimpleShowLayout, TextField, TextInput, RichTextField
+    Create, Datagrid, DateField, DeleteButton, DisabledInput, Edit, EditButton,
+    List, required, Show, ShowButton, SimpleForm, SimpleShowLayout, TextField,
+    TextInput, RichTextField
 } from 'admin-on-rest';
 
+import { cardActionStyle, showActions, listActions } from './defaults.js'
 
-const cardActionStyle = {
-    zIndex: 2,
-    display: 'inline-block',
-    float: 'right'
-};
-
-const DomainListActions = ({ resource, filters, displayedFilters, filterValues, basePath, showFilter }) => (
-    <CardActions style={cardActionStyle}>
-        {filters && React.cloneElement(filters, { resource, showFilter, displayedFilters, filterValues, context: 'button' }) }
-        <CreateButton basePath={basePath} />
-        <RefreshButton />
-    </CardActions>
-);
-
-const DomainShowActions = ({ basePath, data }) => (
-    <CardActions style={cardActionStyle}>
-        <EditButton basePath={basePath} record={data} />
-        <ListButton basePath={basePath} />
-        <DeleteButton basePath={basePath} record={data} />
-        <RefreshButton />
-    </CardActions>
-);
 
 const DomainTitle = ({ record }) => {
     return <span>Domain {record ? `'${record.name}'` : ''}</span>;
 };
 
 export const DomainList = (props) => (
-    <List actions={<DomainListActions />} title='Domains' {...props}>
+    <List actions={<listActions />} title='Domains' {...props}>
         <Datagrid>
             <TextField source='id' />
             <TextField source='name' />
@@ -47,7 +26,7 @@ export const DomainList = (props) => (
 );
 
 export const DomainShow = (props) => (
-    <Show actions={<DomainShowActions />} title={<DomainTitle />} {...props}>
+    <Show actions={<showActions />} title={<DomainTitle />} {...props}>
         <SimpleShowLayout>
             <DateField label='Created' source='created_at' />
             <DateField label='Updated' source='updated_at' />
